requirements.txt
streamlit
pypdf
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
python-dotenv
#InstructorEmbedding 
#sentence_transformers==2.2.2
langchain-community
langchain_openai
langchain_text_splitters
langchain_chroma